# ast: improve String() performance for arrays, sets, objects