--- a/ast/term_bench_test.go
+++ b/ast/term_bench_test.go
@@ -4,6 +4,7 @@
 package ast
 
 import (
+	"encoding/json"
 	"fmt"
 	"strings"
 	"testing"
@@ -102,6 +103,107 @@ func BenchmarkTermHashing(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				_ = s.Hash()
 			}
+		})
+	}
+}
+
+var str string
+var bs []byte
+
+// BenchmarkObjectString generates several objects of different sizes, and
+// marshals them to JSON via two ways:
+//   map[string]int -> ast.Value -> .String()
+// and
+//   map[string]int -> json.Marshal()
+//
+// The difference between these two is relevant for feeding input into the
+// wasm vm: when calling rego.New(...) with rego.Target("wasm"), it's up to
+// the caller to provide the input in parsed form (ast.Value), or
+// raw (interface{}).
+func BenchmarkObjectString(b *testing.B) {
+	var err error
+	sizes := []int{5, 50, 500, 5000}
+
+	for _, n := range sizes {
+		b.Run(fmt.Sprint(n), func(b *testing.B) {
+
+			obj := map[string]int{}
+			for i := 0; i < n; i++ {
+				obj[fmt.Sprint(i)] = i
+			}
+			val := MustInterfaceToValue(obj)
+
+			b.Run("String()", func(b *testing.B) {
+				b.ResetTimer()
+				for i := 0; i < b.N; i++ {
+					str = val.String()
+				}
+			})
+			b.Run("json.Marshal", func(b *testing.B) {
+				b.ResetTimer()
+				for i := 0; i < b.N; i++ {
+					bs, err = json.Marshal(obj)
+					if err != nil {
+						b.Fatal(err)
+					}
+				}
+			})
+		})
+	}
+}
+
+// BenchmarkArrayString compares the performance characteristics of
+// (ast.Value).String() with the stdlib-native json.Marshal. See
+// BenchmarkObjectString above for details.
+func BenchmarkArrayString(b *testing.B) {
+	var err error
+	sizes := []int{5, 50, 500, 5000}
+
+	for _, n := range sizes {
+		b.Run(fmt.Sprint(n), func(b *testing.B) {
+
+			obj := make([]string, n)
+			for i := 0; i < n; i++ {
+				obj[i] = fmt.Sprint(i)
+			}
+			val := MustInterfaceToValue(obj)
+
+			b.Run("String()", func(b *testing.B) {
+				b.ResetTimer()
+				for i := 0; i < b.N; i++ {
+					str = val.String()
+				}
+			})
+			b.Run("json.Marshal", func(b *testing.B) {
+				b.ResetTimer()
+				for i := 0; i < b.N; i++ {
+					bs, err = json.Marshal(obj)
+					if err != nil {
+						b.Fatal(err)
+					}
+				}
+			})
+		})
+	}
+}
+
+func BenchmarkSetString(b *testing.B) {
+	sizes := []int{5, 50, 500, 5000}
+
+	for _, n := range sizes {
+		b.Run(fmt.Sprint(n), func(b *testing.B) {
+
+			val := NewSet()
+			for i := 0; i < n; i++ {
+				val.Add(IntNumberTerm(i))
+			}
+
+			b.Run("String()", func(b *testing.B) {
+				b.ResetTimer()
+				for i := 0; i < b.N; i++ {
+					str = val.String()
+				}
+			})
 		})
 	}
 }